requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
google-generativeai==0.8.3
line-bot-sdk==3.11.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
